requirements/local.txt
Django==4.2.4
djangorestframework==3.14.0
gunicorn==20.1.0
django-redis==5.3.0
celery==5.3.1
redis==4.6.0
timeout-decorator==0.5.0
dask==2023.5.0
